refactor(onboarding): Encapsulate the quickstart experiment in one module
All frontend code the experiment touches now lives in
common/experiments/onboardingQuickstart.ts, including a
useOnboardingQuickstart hook that owns the gate's variant read and
analytics tagging. Cleaning up the experiment is deleting the module
and unpicking its importers.

beep boop
