+++ frontend/common/stores/account-store.js
@@ -4,7 +4,7 @@ import find from 'lodash/find'
 import findIndex from 'lodash/findIndex'
 import get from 'lodash/get'
 import { storageGet, storageSet } from 'common/safeLocalStorage'
-import { clearOnboardingTargetingKey } from 'common/utils/onboardingEntry'
+import { clearOnboardingTargetingKey } from 'common/experiments/onboardingQuickstart'
 import Dispatcher from 'common/dispatcher/dispatcher'
 import BaseStore from './base/_store'
 import data from 'common/data/base/_data'
